# packages/shared-schema/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Wenyi API",
    "description": "Web API and background workers for Wenyi's translation engine.",
    "version": "0.1.dev555+gd99d53be2.d20260923"
  },
  "paths": {
...
    "/projects/{pid}/chapters": {
      "get": {
        "tags": [
          "chapters"
        ],
        "summary": "List Chapters",
        "operationId": "list_chapters_projects__pid__chapters_get",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChapterSummary"
                  },
                  "title": "Response List Chapters Projects  Pid  Chapters Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{pid}/chapters/{ci}": {
      "get": {
        "tags": [
          "chapters"
        ],
        "summary": "Get Chapter",
        "operationId": "get_chapter_projects__pid__chapters__ci__get",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          },
          {
            "name": "ci",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ci"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChapterSegments"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{pid}/chapters/{ci}/title": {
      "put": {
        "tags": [
          "chapters"
        ],
        "summary": "Update Chapter Title",
        "operationId": "update_chapter_title_projects__pid__chapters__ci__title_put",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          },
          {
            "name": "ci",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ci"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChapterTitleUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChapterTitleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{pid}/chapters/{ci}/translate": {
      "post": {
        "tags": [
          "chapters"
        ],
        "summary": "Translate Chapter",
        "operationId": "translate_chapter_projects__pid__chapters__ci__translate_post",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          },
          {
            "name": "ci",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ci"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobEnqueued"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChapterSegments": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "title_translated": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title Translated"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/SegmentOut"
            },
            "type": "array",
            "title": "Segments"
          },
          "review_issues": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Review Issues",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "index",
          "segments"
        ],
        "title": "ChapterSegments"
      },
      "ChapterSummary": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "title_translated": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title Translated"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "pending"
          },
          "word_count": {
            "type": "integer",
            "title": "Word Count",
            "default": 0
          },
          "target_word_count": {
            "type": "integer",
            "title": "Target Word Count",
            "default": 0
          },
          "review_issue_count": {
            "type": "integer",
            "title": "Review Issue Count",
            "default": 0
          },
          "review_status": {
            "type": "string",
            "title": "Review Status",
            "default": "pending"
          }
        },
        "type": "object",
        "required": [
          "index"
        ],
        "title": "ChapterSummary"
      },
      "ChapterTitleOut": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "title_translated": {
            "type": "string",
            "title": "Title Translated"
          }
        },
        "type": "object",
        "required": [
          "index",
          "title_translated"
        ],
        "title": "ChapterTitleOut"
      },
      "ChapterTitleUpdate": {
        "properties": {
          "title_translated": {
            "type": "string",
            "minLength": 1,
            "title": "Title Translated"
          },
          "expected_title_translated": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Title Translated"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title_translated",
          "expected_title_translated"
        ],
        "title": "ChapterTitleUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobEnqueued": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "project_id",
          "kind"
        ],
        "title": "JobEnqueued"
      },
...
      "SegmentOut": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          },
          "display_target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Target"
          },
          "target_before_polish": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Before Polish"
          },
          "anchor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anchor"
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "default": "text"
          }
        },
        "type": "object",
        "required": [
          "index",
          "source"
        ],
        "title": "SegmentOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}